requirements.txt
gradio==4.25.0
requests==2.31.0
python-dotenv==1.0.1
